replace some floats with doubles

+++ macosx/Badger.m
@@ -55,9 +55,9 @@
 
 - (void) updateBadge
 {
-    const float downloadRate = [[NSUserDefaults standardUserDefaults] boolForKey: @"BadgeDownloadRate"]
+    const double downloadRate = [[NSUserDefaults standardUserDefaults] boolForKey: @"BadgeDownloadRate"]
                                 ? tr_sessionGetPieceSpeed_KBps(fLib, TR_DOWN) : 0.0;
-    const float uploadRate = [[NSUserDefaults standardUserDefaults] boolForKey: @"BadgeUploadRate"]
+    const double uploadRate = [[NSUserDefaults standardUserDefaults] boolForKey: @"BadgeUploadRate"]
                                 ? tr_sessionGetPieceSpeed_KBps(fLib, TR_UP) : 0.0;
     
     //only update if the badged values change
